Title screen and temporary audio solution
Code is still a bit messy kinda, but uh we kinda have a title screen. Title screen is temporary tho, it's just the one from the ludum dare prototype.

Need a way to have multiple audio tracks playing at once. For now it's kinda just using the oggplayer from the devkitpro audio template. Audio also need a way to be able to be loaded from the SD card instead of being embedded